requirement.txt
absl-py==0.7.0
aiofiles==0.4.0
aniso8601==3.0.2
astor==0.7.1
astroid==2.1.0
asyncpg==0.18.3
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
Click==7.0
einops==0.1.0
gast==0.2.2
graphene==2.1.3
graphql-core==2.1
graphql-relay==0.4.5
grpcio==1.18.0
h11==0.8.1
h5py==2.9.0
httptools==0.0.11
idna==2.8
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.0
mccabe==0.6.1
mock==2.0.0
numpy==1.16.1
onnx==1.4.1
onnx-tf==1.2.1
pandas==0.24.1
pbr==5.1.2
Pillow==5.4.1
promise==2.2.1
protobuf==3.6.1
pycodestyle==2.5.0
pylint==2.2.2
python-dateutil==2.8.0
python-multipart==0.0.5
pytz==2018.9
PyYAML==3.13
pyzbar==0.1.7
requests==2.21.0
Rx==1.6.1
Shapely==1.6.4.post2
six==1.12.0
SQLAlchemy==1.2.17
starlette==0.10.2
tensorboard==1.12.2
tensorflow==1.13.0rc1
tensorflow-estimator==1.13.0rc0
termcolor==1.1.0
typing==3.6.6
typing-extensions==3.7.2
ujson==1.35
urllib3==1.24.1
uvicorn==0.4.3
uvloop==0.12.1
websockets==7.0
Werkzeug==0.14.1
wrapt==1.11.1
